Rewrote audio sync code.
Audio is now synced with video via timestamps instead of sample rate. The addTime() method is called
with the difference between where the video is and where it should be according to the last audio
timestamp.

The Video class also makes sure that there's always enough frames and audio in the queues to cover
any gap caused by audio/video packet interleaving in the media file.

All audio related functions have been added to the interface IAudioDevice, which Player implements.

The audio sample queue is now also emptied on seeking.